Fixes the unresponsive behavior.
When you include the sfapi swc in, specifically, the app I am trying to
test, it prevents events from propagating from the mouse. This was
traced to the adding of the tool tips to identify elements. I currently
don't know what the fix is but commenting it out fixed the issue.